Create utility functions for getting accessToken and refreshToken and setting browser cookies
