+++ src/app/modules/auth/auth.controller.ts
@@ -6,20 +6,23 @@ import { sendResponse } from "../../utils/sendResponse";
 import httpStatus from "http-status-codes";
 import { AuthServices } from "./auth.service";
 import AppError from "../../errorHelpers/AppError";
+import { setAuthCookie } from "../../utils/setCookie";
 
 const credentialLogin = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
     const loginInfo = await AuthServices.credentialLogin(req.body);
 
-    res.cookie("accessToken", loginInfo.accessToken, {
-      httpOnly: true,
-      secure: false,
-    });
+    // res.cookie("accessToken", loginInfo.accessToken, {
+    //   httpOnly: true,
+    //   secure: false,
+    // });
 
-    res.cookie("refreshToken", loginInfo.refreshToken, {
-      httpOnly: true,
-      secure: false,
-    });
+    // res.cookie("refreshToken", loginInfo.refreshToken, {
+    //   httpOnly: true,
+    //   secure: false,
+    // });
+
+    setAuthCookie(res, loginInfo);
 
     sendResponse(res, {
       success: true,
@@ -32,8 +35,6 @@ const credentialLogin = catchAsync(
 
 const getNewAccessToken = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
-    // const refreshToken = req.cookies.refreshToken;
-    // const tokenInfo = await AuthServices.getNewAccessToken(refreshToken);
     const refreshToken = req.cookies.refreshToken;
 
     if (!refreshToken) {
@@ -47,10 +48,17 @@ const getNewAccessToken = catchAsync(
       refreshToken as string
     );
 
+    // res.cookie("accessToken", tokenInfo.accessToken, {
+    //   httpOnly: true,
+    //   secure: false,
+    // });
+
+    setAuthCookie(res, tokenInfo);
+
     sendResponse(res, {
       success: true,
       statusCode: httpStatus.CREATED,
-      message: "something",
+      message: "User logged in successfully",
       data: tokenInfo,
     });
   }
