+++ src/app/utils/userTokens.ts
@@ -1,6 +1,10 @@
+import { JwtPayload } from "jsonwebtoken";
 import { envVars } from "../config/env";
-import { IUser } from "../modules/user/user.interface";
-import { generateToken } from "./jwt";
+import { IsActive, IUser } from "../modules/user/user.interface";
+import { generateToken, verifyToken } from "./jwt";
+import AppError from "../errorHelpers/AppError";
+import httpStatus from "http-status-codes";
+import { User } from "../modules/user/user.model";
 
 export const createUserTokens = (user: Partial<IUser>) => {
   const jwtPayload = {
@@ -25,4 +29,51 @@ export const createUserTokens = (user: Partial<IUser>) => {
     accessToken,
     refreshToken,
   };
+};
+
+export const createNewAccessTokenWithRefreshToken = async (
+  refreshToken: string
+) => {
+  const verifiedRefreshToken = verifyToken(
+    refreshToken,
+    envVars.JWT_REFRESH_SECRET
+  ) as JwtPayload;
+
+  const isUserExist = await User.findOne({ email: verifiedRefreshToken.email });
+
+  if (!isUserExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, "User does not exist");
+  }
+
+  if (
+    isUserExist.isActive === IsActive.BLOCKED ||
+    isUserExist.isActive === IsActive.INACTIVE
+  ) {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      `User is ${isUserExist.isActive}`
+    );
+  }
+
+  if (isUserExist.isDeleted) {
+    throw new AppError(httpStatus.BAD_REQUEST, "User is deleted");
+  }
+
+  if (!isUserExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, "Email does not exist");
+  }
+
+  const jwtPayload = {
+    userId: isUserExist._id,
+    email: isUserExist.email,
+    role: isUserExist.role,
+  };
+
+  const accessToken = generateToken(
+    jwtPayload,
+    envVars.JWT_ACCESS_SECRET,
+    envVars.JWT_ACCESS_EXPIRES
+  );
+
+  return accessToken;
 };
